test: split provider classification tests
